modularized neural_network to improve readability
